Move bantool functionality to separate script
